+++ src/AuctionService/Entities/Auction.cs
@@ -7,7 +7,7 @@ public class Auction
     public string Seller { get; set; } = null!;
     public string? Winner { get; set; }
 
-    public decimal ReservePrice { get; set; }
+    public int ReservePrice { get; set; }
     public int? SoldAmount { get; set; }
     public int? CurrentHigBid { get; set; }
