+++ src/Contracts/AuctionCreated.cs
@@ -5,7 +5,7 @@ public class AuctionCreated
     public Guid Id { get; set; }
     public string Seller { get; set; }
     public string Winner { get; set; }
-    public decimal ReservePrice { get; set; }
+    public int ReservePrice { get; set; }
     public int SoldAmount { get; set; }
     public int CurrentHigBid { get; set; }
     public DateTime Created { get; set; }
