+++ src/SearchService/Entities/Auction.cs
@@ -6,7 +6,7 @@ public class Auction : Entity
 {
     public string Seller { get; set; } = null!;
     public string? Winner { get; set; }
-    public decimal ReservePrice { get; set; }
+    public int ReservePrice { get; set; }
     public int? SoldAmount { get; set; }
     public int? CurrentHigBid { get; set; }
     public DateTime Created { get; set; }
